change price from int to double data type

+++ Bakery/Models/Order.cs
@@ -6,14 +6,14 @@ namespace Bakery.Models
   {
     public string Title { get; set; }
     public string Description { get; set; }
-    public int Price { get; set; }
+    public double Price { get; set; }
     public string Date { get; set; }
     public string Paid { get; set; }
     public int Id { get; }
     private static int _counter = 1;
     private static List<Order> _orders = new List<Order> {};
 
-    public Order(string title, string description, int price, string date, string paid)
+    public Order(string title, string description, double price, string date, string paid)
     {
       Title = title;
       Description = description;
